Paramaterized delay based on speed

+++ core/src/main/java/edu/cornell/cis3152/team8/Companion.java
@@ -60,7 +60,7 @@ public abstract class Companion extends ObstacleSprite {
 
     private boolean remove;
 
-    private static int MOVE_SPEED = 3;
+    private static float MOVE_SPEED = 3;
 
   private static final float units = 64f;
 
@@ -270,7 +270,7 @@ public abstract class Companion extends ObstacleSprite {
     /**
      * @return companion speed MOVE_SPEED
      */
-    public static int getSpeed(){
+    public static float getSpeed(){
         return MOVE_SPEED;
     }
 
@@ -278,7 +278,7 @@ public abstract class Companion extends ObstacleSprite {
      * Sets companion speed MOVE_SPEED
      * @param speed new speed
      */
-    public static void setSpeed(int speed){
+    public static void setSpeed(float speed){
         MOVE_SPEED = speed;
     }
